+++ CJBCheatsMenu/Framework/Cheats/FarmAndFishing/AutoPetAnimalsCheat.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using System.Linq;
 using CJBCheatsMenu.Framework.Components;
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
@@ -8,8 +7,8 @@ using StardewValley.Menus;
 
 namespace CJBCheatsMenu.Framework.Cheats.FarmAndFishing
 {
-    /// <summary>A cheat which prevents NPC friendships from decaying.</summary>
-    internal class NoAnimalFriendshipDecayCheat : BaseCheat
+    /// <summary>A cheat which automatically pets farm animals.</summary>
+    internal class AutoPetAnimalsCheat : BaseCheat
     {
         /*********
         ** Fields
@@ -22,9 +21,9 @@ namespace CJBCheatsMenu.Framework.Cheats.FarmAndFishing
         public override IEnumerable<OptionsElement> GetFields(CheatContext context)
         {
             yield return new CheatsOptionsCheckbox(
-                label: I18n.Farm_NoDecay(),
-                value: context.Config.NoAnimalFriendshipDecay,
-                setValue: value => context.Config.NoAnimalFriendshipDecay = value
+                label: I18n.Farm_AutoPetAnimals(),
+                value: context.Config.AutoPetAnimals,
+                setValue: value => context.Config.AutoPetAnimals = value
             );
         }
 
@@ -33,25 +32,24 @@ namespace CJBCheatsMenu.Framework.Cheats.FarmAndFishing
         /// <param name="needsUpdate">Whether the cheat should be notified of game updates.</param>
         /// <param name="needsInput">Whether the cheat should be notified of button presses.</param>
         /// <param name="needsRendering">Whether the cheat should be notified of render ticks.</param>
-        public override void OnConfig(CheatContext context, out bool needsInput, out bool needsUpdate,
-            out bool needsRendering)
+        public override void OnConfig(CheatContext context, out bool needsInput, out bool needsUpdate, out bool needsRendering)
         {
             needsInput = false;
-            needsUpdate = context.Config.NoAnimalFriendshipDecay;
+            needsUpdate = context.Config.AutoPetAnimals;
             needsRendering = false;
         }
 
         /// <summary>Handle the player loading a save file.</summary>
         /// <param name="context">The cheat context.</param>
         /// <summary>Handle a game update if <see cref="ICheat.OnSaveLoaded"/> indicated updates were needed.</summary>
-        /// <param name="context">The cheat context.</param>
         /// <param name="e">The update event arguments.</param>
         public override void OnUpdated(CheatContext context, UpdateTickedEventArgs e)
         {
             if (!e.IsOneSecond || !Context.IsWorldReady)
                 return;
-            IEnumerable<FarmAnimal> animals = Game1.getFarm().getAllFarmAnimals().Distinct();
-            foreach (FarmAnimal animal in animals) animal.wasPet.Set(true);
+
+            foreach (FarmAnimal animal in Game1.getFarm().getAllFarmAnimals())
+                animal.wasPet.Value = true;
         }
     }
 }